+++ libndtypes/ndtypes.h
@@ -427,10 +427,10 @@ struct _ndt {
 
             struct {
                 int64_t itemsize;
-                int64_t start;
-                int64_t stop;
-                int64_t step;
-                int64_t noffsets;
+                int32_t start;
+                int32_t stop;
+                int32_t step;
+                int32_t noffsets;
                 const int32_t *offsets;
             } VarDim;
 
@@ -578,7 +578,7 @@ NDTYPES_API ndt_t *ndt_any_kind(ndt_context_t *ctx);
 NDTYPES_API ndt_t *ndt_fixed_dim(int64_t shape, ndt_t *type, char order, ndt_context_t *ctx);
 NDTYPES_API ndt_t *ndt_symbolic_dim(char *name, ndt_t *type, ndt_context_t *ctx);
 NDTYPES_API ndt_t *ndt_var_dim(ndt_t *type, bool copy_offsets, int32_t noffsets, const int32_t *offsets,
-                               int64_t start, int64_t stop, int64_t step, ndt_context_t *ctx);
+                               int32_t start, int32_t stop, int32_t step, ndt_context_t *ctx);
 NDTYPES_API ndt_t *ndt_ellipsis_dim(char *name, ndt_t *type, ndt_context_t *ctx);
 
 NDTYPES_API ndt_t *ndt_option(ndt_t *type, ndt_context_t *ctx);
